Fixed a bug that made the client hang
The tags retrieval for guile code uses wasn't very good. It turned out
that the functions that were provided by MOCP weren't suited for that
kind of stuff. They updated the playlist and we wanted just to receive
tags from the server (even when the song wasn't on the playlist) or
from the local playlist.

The new function - guile_get_tags_from_srv - isn't tested for the
retrieval of the tags from the server yet. In normal circumstances it
wouldn't be necessary because we want to get the tags for the current
song which should be present already.