Added queue for events coming from guile + some improvements
The queue is necessary because guile code can be asynchronous. If the
user wanted to interact with the interface or any other part of the
client it's necessary to provide a way that is thread safe for that.

The guile event queue is a scheme list with an additional tail field.
The list stores the SCM thunks that can be evaluated later inside the
interface_loop().

Other stuff added is convenience macros for asserting that the
client-specific code runs only on the client and likewise for the
server.