Wake up on Guile event
It works like in the server: there is a wake up pipe that is written
to when a Guile event is added to the queue. It's used to wake up the
pselect() in interface_loop(). The difference is that the additional
variable is used as a flag if the wake_up_pipe was already written to
before the value was consumed. It will prevent unnecessary wake-ups if
several events are pushed at the same time.